--- src/lib/colors/formatRgb/index.js
@@ -1,7 +1,0 @@
-/**
- * Formats an rgba into a css rgb/a property
- * @returns {String}
- */
-const formatRgb = ({ r, g, b, a = 1 }) => (a !== 1 ? `rgba(${r}, ${g}, ${b}, ${a})` : `rgb(${r}, ${g}, ${b})`);
-
-export default formatRgb;
