+++ src/lib/colors/hexToRgb/index.ts
@@ -1,15 +1,17 @@
+import { Color } from '../../types';
+
 export const hexRegExp = new RegExp(/^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i);
 export const sortHexRegExp = new RegExp(/^#?([a-f\d])([a-f\d])([a-f\d])$/i);
 
-const formatHexNumber = item => parseInt(item.repeat(3 - item.length), 16);
+const formatHexNumber = (item: string): number => parseInt(item.repeat(3 - item.length), 16);
 
 /**
  * Converts a hex color string into rgb
- * @param {String} hex - Hex color string
- * @param {Number} opacity - The opacity to use
- * @returns {Object}
+ * @param hex - Hex color string
+ * @param opacity - The opacity to use
+ * @returns RGB color object or undefined if invalid hex
  */
-const hexToRgb = (hex, opacity = 1) => {
+const hexToRgb = (hex: string, opacity: number = 1): Color | undefined => {
   const result = hexRegExp.exec(hex) || sortHexRegExp.exec(hex);
 
   return result
@@ -22,4 +24,4 @@ const hexToRgb = (hex, opacity = 1) => {
     : undefined;
 };
 
-export default hexToRgb;
+export default hexToRgb;
