--- src/lib/index.js
@@ -1,7 +1,0 @@
-export { default as sortPalette } from './sortPalette';
-export { default as noop } from './noop';
-export { default as clampAngle } from './clampAngle';
-export { default as centerOffset } from './centerOffset';
-export { default as pointDegrees } from './pointDegrees';
-export { default as snapAngle } from './snapAngle';
-export { default as getGradientPreview } from './getGradientPreview';
